pyproject.toml
[project]
name = "impatient"
version = "1.7.0"
description = "IMPatienT: an integrated web application to digitize, process and explore multimodal patient data"
authors = [{ name = "Corentin Meyer", email = "contact@cmeyer.fr" }]
maintainers = [
    { name = "Corentin Meyer", email = "contact@cmeyer.fr" }
]
requires-python = ">= 3.10"

dependencies = [
    "scikit-learn<2",
    "scikit-image<1",
    "dash<4",
    "Flask<4",
    "Flask-WTF<2",
    "WTForms<3",
    "wtforms-validators<2",
    "Flask-SQLAlchemy<4",
    "Flask-Mail<1",
    "python-dotenv<1",
    "pandas<3",
    "numpy<2",
    "plotly<7",
    "Pillow<12",
    "joblib<2",
    "dash-core-components<3",
    "dash-bootstrap-components<2",
    "Flask-Migrate<5",
    "Flask-Session<1",
    "pytesseract<1",
    "pdf2image<2",
    "spacy<=3.8.7",
    "thefuzz<1",
    "email-validator<3",
    "dash-html-components<3",
    "gunicorn<24",
    "opencv-python-headless<5",
    "python-Levenshtein<0.27.1",
    "PyJWT<2.4.0",
    "en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl",
    "Flask-Cors<7",
    "textacy<1",
    "bleach<5.1.1",
    "pronto<2.6.5",
    "jsonschema<4.20.0",
    "cairosvg<3.0.0",
    "huggingface-hub>=0.33.0",
]

[dependency-groups]
dev = [
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/impatient"]
